# ss/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movies-app-ss",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/banners": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "Get list of banners.",
                "operationId": "b9d60a753393ddfebc34ce9a774b5a72",
                "responses": {
                    "200": {
                        "description": "Display a listing of banners."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Banner"
                ],
                "summary": "Create new banner.",
                "operationId": "7d0921a177e15da085d65b1ab420d92b",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of banner",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/name"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Path of image.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/image"
                        }
                    },
                    {
                        "name": "banner_category",
                        "in": "query",
                        "description": "Category.",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/banner_category"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Banner"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "api/banners/{id}": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "Get banner by id",
                "operationId": "1796ba5085ae905013950f5542df8dbe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of films."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Banner"
                ],
                "summary": "Update banner by id.",
                "operationId": "af2c6a0f5142c0cbb84b85e24e1624d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/id"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of banner",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/name"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Link of banner.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/image"
                        }
                    },
                    {
                        "name": "banner_category",
                        "in": "query",
                        "description": "Category.",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/banner_category"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Banner"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Banner"
                ],
                "summary": "Delete banner by id.",
                "operationId": "d588badf7804dad9bbc57a6bb556c7c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Banner/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "api/configurations": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Get list of configurations.",
                "operationId": "56c4593b5b5e67fcb2f4857348883a5f",
                "responses": {
                    "200": {
                        "description": "Display a listing of configurations."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Create new configuration.",
                "operationId": "9a786befc915cbd212eec152de2a07e1",
                "parameters": [
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "Status of category.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Configuration/properties/domain"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Configuration"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "api/configurations/{id}": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Get configuration by id",
                "operationId": "b72937c09e5553be6dd5a90d1c0d7ac4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Configuration/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of films."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Update configuration by id.",
                "operationId": "105e288ee15341025b9f14642957e05e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Configuration/properties/id"
                        }
                    },
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "Domain.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Configuration/properties/domain"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Configuration"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Delete configuration by id.",
                "operationId": "36bee470c2933fbb9091dfcb574711c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Configuration/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Banner": {
                "required": [
                    "name",
                    "image"
                ],
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/BaseModel/properties/id"
                    },
                    "name": {
                        "type": "string",
                        "example": "Banner 1"
                    },
                    "image": {
                        "type": "string",
                        "example": "upload/image.png"
                    },
                    "banner_category": {
                        "type": "integer",
                        "example": "1"
                    },
                    "status": {
                        "type": "integer",
                        "example": "Trạng thái: 1-kích hoạt, 2-vô hiệu, 3-đã xóa"
                    },
                    "position": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_by"
                    },
                    "updated_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "update_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Banner"
                }
            },
...
            "Configuration": {
                "required": [
                    "domain"
                ],
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/BaseModel/properties/id"
                    },
                    "domain": {
                        "type": "string"
                    },
                    "domain_template_id": {
                        "type": "integer"
                    },
                    "site_name": {
                        "type": "string"
                    },
                    "site_description": {
                        "type": "string"
                    },
                    "keyword": {
                        "type": "string"
                    },
                    "company_name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "header": {
                        "type": "string"
                    },
                    "footer": {
                        "type": "string"
                    },
                    "logo": {
                        "type": "string"
                    },
                    "facebook": {
                        "type": "string"
                    },
                    "youtube": {
                        "type": "string"
                    },
                    "instagram": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "example": "Trạng thái: 1-kích hoạt, 2-vô hiệu, 3-đã xóa"
                    },
                    "position": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_by"
                    },
                    "updated_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "update_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Configuration"
                }
            },
...
        }
    }
}